pyproject.toml
[project]
name = "reelm"
version = "0.2.0"
description = "Your personal media agent — AI-native alternative to Sonarr/Radarr"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.0",
    "uvicorn>=0.40.0",
    "httpx>=0.28.1",
    "transmission-rpc>=8.0.0a4",
    "tmdbsimple>=2.9.1",
    "xmltodict>=1.0.2",
    "jmespath>=1.1.0",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.2.1",
    "anyascii>=0.3.3",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.0",
    "ty>=0.0.1a7",
    "pip-audit>=2.9.0",
    "yamllint>=1.37.0",
    "diff-cover>=9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcps"]

[tool.uv]
prerelease = "allow"

[tool.ruff]
target-version = "py312"
line-length = 140
output-format = "concise"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "S", "SIM", "PTH", "RUF"]
ignore = [
    "SIM108",  # ternary — often hurts readability
    "SIM110",  # any() rewrite — often hurts readability
    "RUF001",  # ambiguous unicode — we use Cyrillic intentionally (Russian alt titles)
    "RUF015",  # next(iter()) vs [0] — readability preference
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106", "E402"]  # assert in tests; hardcoded test creds; env setup before imports

[tool.ty.environment]
python-version = "3.12"
root = ["src"]

[tool.ty.rules]
possibly-unresolved-reference = "error"
invalid-argument-type = "error"
missing-argument = "error"
unsupported-operator = "error"
division-by-zero = "error"
unused-ignore-comment = "warn"
redundant-cast = "warn"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers --cov --cov-report=term:skip-covered --cov-report=xml --cov-fail-under=90"
pythonpath = ["src"]
markers = [
    "unit: unit tests",
]

[tool.coverage.run]
source = ["src/mcps"]
branch = true

[tool.coverage.report]
show_missing = false
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "\\.\\.\\.",
]
